requirements.txt
# OpenAI SDK
openai>=1.0.0

# Pydantic for data validation
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP client for async operations
aiohttp>=3.8.0

# Google Generative AI SDK (Gemini)
# Note: google-generativeai package has been deprecated, migrate to google-genai
google-generativeai>=0.3.0

# Async support
asyncio-mqtt>=0.10.0; python_version >= '3.7'

# Python standard library additional dependencies (optional, for better development experience)
pathlib2>=2.3.7; python_version < '3.4'

# Development and testing tools (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.7.0

# Documentation generation (optional)
# mkdocs>=1.5.0
# mkdocs-material>=9.0.0

# Rich for enhanced output (optional)
# rich>=13.0.0